# Also compute a `unique` count based on the user IP