--- a/src/file.rs
+++ b/src/file.rs
@@ -1,19 +1,18 @@
 use flate2::read::GzDecoder;
-use std::error::Error;
 use std::ffi::OsStr;
 use std::fs::File;
 use std::io::BufRead;
 use std::io::BufReader;
 use std::path::Path;
 
-pub fn reader(filename: &str, opts: &super::Options) -> Box<BufRead> {
+pub fn reader(filename: &str, opts: &super::Options) -> Box<dyn BufRead> {
   if opts.verbose {
     println!("Opening log file {}", filename);
   }
 
   let path = Path::new(filename);
   let file = match File::open(&path) {
-    Err(why) => panic!("couldn't open {}: {}", path.display(), why.description()),
+    Err(why) => panic!("couldn't open {}: {}", path.display(), why.to_string()),
     Ok(file) => file,
   };