Based of: https://github.com/rubytogether/kirby/pull/19

### Feature

As discussed on Bundler's slack. The usage statistics are not very usable, because a small minority of users can generate lots of traffic and trust the top of the stats.

This for instance happens with large fleets of AWS EC2 instances running older Linux distributions with the system Ruby 2.3, causing 2.3 to be almost half of all requests.

To counteract this, we can for each metric only count an IP once. Meaning some kind of automated script doing 1k requests/day will count for 1, just like a human user doing 5 requests/day.

The "unicity key" is `[user_agent, client_ip, day]`, so a single IP doing requests with multiple ruby or bundler versions will be counted once per day in the respective buckets.

In essence it's the same idea than "daily unique visitor" kind of metrics.

This should hopefully give a more relevant picture of Ruby and Bundler version usage.

### Implementation

Instead of replacing the existing count, I chose to count both, the downside is that it changes the result format. It's easy to just replace it if we want, I just didn't want to commit to either solution at this stage.

Also `kirby` panic quite quickly on my machine, but it already does so on `main`:

```
thread '<unnamed>' panicked at 'attempted to leave type `nodrop::imp::NoDrop<(epoch::Epoch, garbage::Bag)>` uninitialized, which is invalid', /Users/byroot/.cargo/registry/src/github.com-1ecc6299db9ec823/crossbeam-epoch-0.3.1/src/sync/queue.rs:57:28
stack backtrace:
```

I have no idea what might cause this and how to fix it.


However It sill run long enough that I was able to test the feature:
```
rustup run nightly cargo run test/sample_10_dups.log 
{
  "files":[
    "test/sample_10_dups.log"
  ],
  "ran_at":"2022-02-03T12:47:46Z",
  "stats":{
    "2018-04-16":{
      "bundler":{
        "1.16.1":{
          "total":2,
          "unique":1
        }
      },
      "ci":{
        
      },
      "gemstash":{
        
      },
      "platform":{
        "x86_64-linux":{
          "total":2,
          "unique":1
        },
        "x86_64-pc-linux-gnu":{
          "total":2,
          "unique":1
        }
      },
      "ruby":{
        "2.1.8":{
          "total":2,
          "unique":1
        },
        "2.4.1":{
          "total":2,
          "unique":1
        }
      },
      "rubygems":{
        "2.2.5":{
          "total":2,
          "unique":1
        },
        "2.6.11":{
          "total":2,
          "unique":1
        }
      },
      "server_region":{
        
      },
      "tls_cipher":{
        "ECDHE-RSA-AES128-GCM-SHA256":{
          "total":4,
          "unique":2
        }
      }
    }
  }
}
```

### NB

This is my first time writing rust ever so I have no clue what I'm doing.