@indirect 2022-02-08
I think changing the output format is fine, it's just one additional key that can be read by the background job in rubytogether/ecosystem 👍🏻 looks good to me!