@casperisfine 2022-02-08
Yeah, I'll prepare a PR to add a column in ecosystem and update the import script.